requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
httpx==0.26.0
quart==0.17.0
werkzeug==2.0.3
pydantic==1.10.13
fastapi==0.109.0